ml_service/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pandas
pydantic
redis
requests

torch
torchvision
opencv-python-headless
pillow
